+++ api/src/org/labkey/api/study/assay/ParticipantVisitResolver.java
@@ -24,7 +24,7 @@ import org.labkey.api.study.ParticipantVisit;
 import java.util.Date;
 
 /**
- * Helps resolve and validate information during an assay's copy-to-study operation. Since each row will need to be
+ * Helps resolve and validate information during an assay's link to study operation. Since each row will need to be
  * associated with a participant and date/visit to be incorporated into the target study dataset, different
  * implementations may use other context to automate filling them in. For example, an implementation could use a
  * specimen ID column that's in the assay data to find the associated specimen in the target study and automatically
