+++ api/src/org/labkey/api/study/assay/RunDatasetContextualRoles.java
@@ -54,7 +54,7 @@ public class RunDatasetContextualRoles implements HasContextualRoles
     /**
      * Returns a contextual ReaderRole if the user has permission to
      * <b>at least one of</b> the study datasets that the run results have
-     * been copied to.
+     * been linked to.
      *
      * @return a singleton ReaderRole set or null
      */
@@ -80,7 +80,7 @@ public class RunDatasetContextualRoles implements HasContextualRoles
     /**
      * Returns a contextual ReaderRole if the user has permission to
      * <b>at least one of</b> the study datasets that the run results have
-     * been copied to.
+     * been linked to.
      *
      * @param container the container
      * @param user the user
@@ -129,7 +129,7 @@ public class RunDatasetContextualRoles implements HasContextualRoles
                 datasetColumnNames.add(columnName);
         }
 
-        // table contains no dataset columns if results haven't been copied
+        // table contains no dataset columns if results haven't been linked
         if (datasetColumnNames.size() == 0)
             return null;
