+++ experiment/src/org/labkey/experiment/XarReader.java
@@ -1620,7 +1620,7 @@ public class XarReader extends AbstractXarImporter
                 value = LsidUtils.resolveLsidFromTemplate(value, getRootContext());
             }
 
-            if (StudyPublishService.AUTO_COPY_TARGET_PROPERTY_URI.equals(simpleProp.getOntologyEntryURI()) && value != null)
+            if (StudyPublishService.AUTO_LINK_TARGET_PROPERTY_URI.equals(simpleProp.getOntologyEntryURI()) && value != null)
             {
                 Container autoCopyContainer = ContainerManager.getForPath(value);
                 if (autoCopyContainer != null)
