+++ experiment/src/org/labkey/experiment/controllers/exp/ExperimentController.java
@@ -716,7 +716,7 @@ public class ExperimentController extends SpringActionController
             }
 
             ActionURL linkToStudyHistoryURL = new ActionURL(); // Rosaline: TODO in LinkToStudyAction story
-            ActionButton linkToStudyHistoryButton = new ActionButton(linkToStudyHistoryURL, "Link-to-Study History", ActionButton.Action.LINK);
+            ActionButton linkToStudyHistoryButton = new ActionButton(linkToStudyHistoryURL, "Link to Study History", ActionButton.Action.LINK);
             linkToStudyHistoryButton.setDisplayPermission(InsertPermission.class);
             detailsView.getDataRegion().getButtonBar(DataRegion.MODE_DETAILS).add(linkToStudyHistoryButton);
