+++ study/src/org/labkey/study/query/PublishedRecordQueryView.java
@@ -97,7 +97,7 @@ public class PublishedRecordQueryView extends DatasetQueryView
     private class PublishedRecordDataRegion extends DataRegion
     {
         private static final String MISSING_ROWS_MSG = "%s rows that were previously linked in this event have been recalled (or deleted)." +
-                " The audit record(s) of the deleted rows can be found in the <a href=\"%s\">link-to-study history view</a>, or the" +
+                " The audit record(s) of the deleted rows can be found in the <a href=\"%s\">link to study history view</a>, or the" +
                 " study dataset history view.";
 
         private int _count;
