+++ study/test/src/org/labkey/test/tests/study/AssayTest.java
@@ -511,7 +511,7 @@ public class AssayTest extends AbstractAssayTest
         //select all the data rows and click publish
         DataRegionTable table = new DataRegionTable("Data", this);
         table.checkAllOnPage();
-        table.clickHeaderButtonAndWait("Copy to Study");
+        table.clickHeaderButtonAndWait("Link to Study");
 
         //the target study selected before was Study2, but the PI is not an editor there
         //so ensure that system has correctly caught this fact and now asks the PI to
@@ -519,7 +519,7 @@ public class AssayTest extends AbstractAssayTest
         //an editor
 
         //ensure warning
-        assertTextPresent("WARNING: You do not have permissions to copy to one or more of the selected run's associated studies.");
+        assertTextPresent("WARNING: You do not have permissions to link to one or more of the selected run's associated studies.");
 
         //ensure that Study2 and Study 3 are not available in the target study drop down
         assertElementNotPresent(Locator.xpath("//select[@name='targetStudy']/option[.='" +
@@ -535,10 +535,10 @@ public class AssayTest extends AbstractAssayTest
         selectOptionByText(Locator.xpath("//select[@name='targetStudy']"), getTargetStudyOptionText(TEST_ASSAY_FLDR_STUDY1));
 
         clickButton("Next");
-        assertTextPresent("Copy to " + TEST_ASSAY_FLDR_STUDY1 + " Study: Verify Results");
+        assertTextPresent("Link to " + TEST_ASSAY_FLDR_STUDY1 + " Study: Verify Results");
 
         setFormElement(Locator.name("visitId"), "301.5");
-        clickButton("Copy to Study");
+        clickButton("Link to Study");
 
         log("Verifying that the data was published");
         _customizeViewsHelper.openCustomizeViewPanel();
@@ -585,20 +585,20 @@ public class AssayTest extends AbstractAssayTest
 
         // Set a filter so that we know we're recalling SecondRun
         DataRegionTable region = new DataRegionTable("query", this);
-        region.setFilter("Comment", "Starts With", "3 row(s) were copied to a study from the assay");
+        region.setFilter("Comment", "Starts With", "3 row(s) were linked to a study from the assay");
         doAndWaitForPageToLoad(() -> region.detailsLink(region.getRowIndex("Assay/Protocol", TEST_ASSAY)).click());
 
-        DataRegionTable copyStudy = new DataRegionTable("Dataset", this);
-        copyStudy.checkAll();
+        DataRegionTable linkStudy = new DataRegionTable("Dataset", this);
+        linkStudy.checkAll();
         doAndWaitForPageToLoad(() ->
         {
-            copyStudy.clickHeaderButton("Recall Rows");
+            linkStudy.clickHeaderButton("Recall Rows");
             acceptAlert();
         });
         assertTextPresent("row(s) were recalled to the assay: " + TEST_ASSAY);
 
-        // Set a filter so that we know we're looking at the copy event for SecondRun again
-        region.setFilter("Comment", "Starts With", "3 row(s) were copied to a study from the assay");
+        // Set a filter so that we know we're looking at the link event for SecondRun again
+        region.setFilter("Comment", "Starts With", "3 row(s) were linked to a study from the assay");
 
         // verify audit entry was adjusted
         doAndWaitForPageToLoad(() -> region.detailsLink(region.getRowIndex("Assay/Protocol", TEST_ASSAY)).click());
@@ -608,9 +608,9 @@ public class AssayTest extends AbstractAssayTest
     }
 
     /**
-     * Designed to test automatic timepoint generation when copying to a date based study.
+     * Designed to test automatic timepoint generation when linking to a date based study.
      * Most tests of timepoint matching are covered by separate junit tests; however,
-     * this will create 1 pre-existing timepoint, and when copying data this timepoint should be
+     * this will create 1 pre-existing timepoint, and when linking data this timepoint should be
      * chosen for appropriate records.
      */
     @LogMethod
@@ -639,7 +639,7 @@ public class AssayTest extends AbstractAssayTest
         //select all the data rows and click publish
         DataRegionTable table = new DataRegionTable("Data", getDriver());
         table.checkAll();
-        table.clickHeaderButtonAndWait("Copy to Study");
+        table.clickHeaderButtonAndWait("Link to Study");
 
         checkCheckbox(Locator.xpath("//input[@id='chooseStudy']"));
 
@@ -647,7 +647,7 @@ public class AssayTest extends AbstractAssayTest
         selectOptionByText(Locator.xpath("//select[@name='targetStudy']"), getTargetStudyOptionText(TEST_ASSAY_FLDR_STUDY3));
 
         clickButton("Next");
-        assertTextPresent("Copy to " + TEST_ASSAY_FLDR_STUDY3 + " Study: Verify Results");
+        assertTextPresent("Link to " + TEST_ASSAY_FLDR_STUDY3 + " Study: Verify Results");
 
         //populate initial set of values and verify the timepoint preview column
         String[] dates = new String[]{"2000-02-02", "2000-03-03", "2000-04-04", "2000-05-05", "2000-06-06", "2001-01-01", "2000-01-01", "2000-02-02", "2000-03-03"};
@@ -663,8 +663,8 @@ public class AssayTest extends AbstractAssayTest
         setFormElement(Locator.xpath("(//input[@name='participantId'])[3]"), "new3");
         setFormElement(Locator.xpath("(//input[@name='participantId'])[4]"), "new4");
 
-        DataRegionTable copyStudy = new DataRegionTable("Data", getDriver());
-        copyStudy.clickHeaderButtonAndWait("Re-Validate");
+        DataRegionTable linkStudy = new DataRegionTable("Data", getDriver());
+        linkStudy.clickHeaderButtonAndWait("Re-Validate");
 
         //validate timepoints:
         assertElementPresent(Locator.xpath("//td[text()='Day 32 - 39' and following-sibling::td/a[text()='AAA07XMC-02'] and following-sibling::td[text()='301.0']]"));
@@ -677,7 +677,7 @@ public class AssayTest extends AbstractAssayTest
         assertElementPresent(Locator.xpath("//td[text()='Day 32 - 39' and following-sibling::td/a[text()='BAQ00051-08'] and following-sibling::td[text()='8.0']]"));
         assertElementPresent(Locator.xpath("//td[text()='Preexisting Timepoint' and following-sibling::td/a[text()='BAQ00051-11'] and following-sibling::td[text()='9.0']]"));
 
-        copyStudy.clickHeaderButtonAndWait("Copy to Study");
+        linkStudy.clickHeaderButtonAndWait("Link to Study");
 
         log("Verifying that the data was published");
         assertTextPresent(
@@ -709,9 +709,9 @@ public class AssayTest extends AbstractAssayTest
 
 
     /**
-     * Designed to test automatic timepoint generation when copying to a date based study.
+     * Designed to test automatic timepoint generation when linking to a date based study.
      * Most tests of timepoint matching are covered by separate junit tests; however,
-     * this will create 1 pre-existing timepoint, and when copying data this timepoint should be
+     * this will create 1 pre-existing timepoint, and when linking data this timepoint should be
      * chosen for appropriate records.
      */
     @LogMethod
@@ -740,7 +740,7 @@ public class AssayTest extends AbstractAssayTest
         //select all the data rows and click publish
         DataRegionTable table = new DataRegionTable("Data", getDriver());
         table.checkAll();
-        table.clickHeaderButtonAndWait("Copy to Study");
+        table.clickHeaderButtonAndWait("Link to Study");
 
         checkCheckbox(Locator.xpath("//input[@id='chooseStudy']"));
 
@@ -748,7 +748,7 @@ public class AssayTest extends AbstractAssayTest
         selectOptionByText(Locator.xpath("//select[@name='targetStudy']"), getTargetStudyOptionText(TEST_ASSAY_FLDR_STUDY2));
 
         clickButton("Next");
-        assertTextPresent("Copy to " + TEST_ASSAY_FLDR_STUDY2 + " Study: Verify Results");
+        assertTextPresent("Link to " + TEST_ASSAY_FLDR_STUDY2 + " Study: Verify Results");
 
         //populate initial set of values and verify the timepoint preview column
         String[] visits = new String[]{"302", "33", "4", "70"};
@@ -764,8 +764,8 @@ public class AssayTest extends AbstractAssayTest
         setFormElement(Locator.xpath("(//input[@name='participantId'])[3]"), "new3");
         setFormElement(Locator.xpath("(//input[@name='participantId'])[4]"), "new4");
 
-        DataRegionTable copyStudy = new DataRegionTable("Data", getDriver());
-        copyStudy.clickHeaderButtonAndWait("Re-Validate");
+        DataRegionTable linkStudy = new DataRegionTable("Data", getDriver());
+        linkStudy.clickHeaderButtonAndWait("Re-Validate");
 
         //validate timepoints:
         assertElementPresent(Locator.xpath("//td[text()='Test Visit3' and following-sibling::td/a[text()='AAA07XMC-02']]"));
@@ -778,7 +778,7 @@ public class AssayTest extends AbstractAssayTest
         assertElementPresent(Locator.xpath("//td[text()='Test Visit1' and following-sibling::td/a[text()='BAQ00051-08']]"));
         assertElementPresent(Locator.xpath("//td[text()='Test Visit1' and following-sibling::td/a[text()='BAQ00051-11']]"));
 
-        copyStudy.clickHeaderButtonAndWait("Copy to Study");
+        linkStudy.clickHeaderButtonAndWait("Link to Study");
 
         log("Verifying that the data was published");
         assertTextPresent(
@@ -833,7 +833,7 @@ public class AssayTest extends AbstractAssayTest
         assertTextPresent(TEST_ASSAY_DATA_PROP_NAME + "edit");
         assertTextNotPresent(TEST_ASSAY_DATA_PROP_NAME + 4);
 
-        AuditLogTest.verifyAuditEvent(this, AuditLogTest.ASSAY_AUDIT_EVENT, AuditLogTest.COMMENT_COLUMN, "were copied to a study from the assay: " + TEST_ASSAY, 5);
+        AuditLogTest.verifyAuditEvent(this, AuditLogTest.ASSAY_AUDIT_EVENT, AuditLogTest.COMMENT_COLUMN, "were linked to a study from the assay: " + TEST_ASSAY, 5);
     }
 
     @LogMethod
@@ -894,16 +894,16 @@ public class AssayTest extends AbstractAssayTest
         region.clearAllFilters("SpecimenID");
         verifySpecimensPresent(3, 2, 3);
 
-        // Verify that the correct copied to study column is present
-        assertTextPresent("Copied to Study 1 Study");
+        // Verify that the correct linked to study column is present
+        assertTextPresent("Linked to Study 1 Study");
 
-        log("Testing copy to study availability");
+        log("Testing link to study availability");
         clickProject(getProjectName());
         clickAndWait(Locator.linkWithText(TEST_RUN3));
 
         region = new DataRegionTable("Data", this);
         region.checkAll();
-        region.clickHeaderButtonAndWait("Copy to Study");
+        region.clickHeaderButtonAndWait("Link to Study");
         clickButton("Next");
 
         verifySpecimensPresent(0, 0, 3);
